Start Scene and End Scene Added
I added to the 2 scenes for ending and start, we can decide on designing it more at a later point, so it isn't just a GUI with no environment. I did a bit of an animation for the end scene, so it looks nice, meaning that i also created an animations folder, so if you make any animations, make sure to put them there.